server/requirements.txt
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
fastapi
uvicorn
langchain-community
